Moved unique name check to the WorkflowValidator
